Side-project while looking at Machine Learning
While I was learning about SVMs, as I was trying to implement the algorithm myself, I tried to devise a method of iterating through all rotations to find the best classifier plane. An hour of research later, this little piece of code was born - given one radius, and n-1 angles, it'll calculate the plane equation which fits those classifications. Unfortunately, I didn't have the time to look indepth into the maths behind why it works the way it works - the underlying process is just based off the patterns that I've observed in the formula to calculate the position vector given a radius and n-1 angles.
Fun, but not completely Explored - 6/10