Code written for UCL COMP102 Theory Assignment 5
Uses pathfinding to determine the correct series of steps required to achieve the win condition in the minimum amount of steps in a game.
The game consists of a 4x4 grid, where on each square there can exist either a mole or not - if there is a mole, the mole can be whacked, removing the mole and all moles in a plus shape around it - however if there are any squares in the plus which do not contain moles, following the whack, moles will appear there. The aim is to remove all moles.